seperate AgeWarning from WithAgeWarning component
