+++ apps-rendering/src/components/WithAgeWarning/index.tsx
@@ -3,22 +3,11 @@ import { css } from '@emotion/react';
 import { background } from '@guardian/common-rendering/src/editorialPalette';
 import type { Tag } from '@guardian/content-api-models/v1/tag';
 import type { ArticleFormat } from '@guardian/libs';
-import {
-	from,
-	palette,
-	textSans,
-	visuallyHidden,
-} from '@guardian/source-foundations';
 import type { Option } from '@guardian/types';
 import { OptionKind } from '@guardian/types';
+import { AgeWarning } from 'components/AgeWarning';
 import { isComment, isNews } from 'item';
 import { articleWidthStyles, darkModeCss } from 'styles';
-
-type AgeWarningProps = {
-	age: string;
-	isScreenReader?: boolean;
-	size?: 'small' | 'medium';
-};
 
 interface WithAgeWarningProps {
 	tags: Tag[];
@@ -66,58 +55,6 @@ const getAgeWarning = (
 		}
 	}
 	return message;
-};
-
-const ageWarningStyles = (isSmall: boolean): SerializedStyles => css`
-	${isSmall ? textSans.xxsmall() : textSans.medium()};
-	color: ${palette.neutral[7]};
-	background-color: ${palette.brandAlt[400]};
-	display: inline-block;
-
-	> strong {
-		font-weight: bold;
-	}
-
-	padding: ${isSmall ? '3px 5px' : '6px 10px'};
-
-	${from.mobileLandscape} {
-		padding-left: ${isSmall ? '6px' : '12px'};
-	}
-
-	${from.leftCol} {
-		padding-left: ${isSmall ? '5px' : '10px'};
-	}
-`;
-
-const ageWarningScreenReader = css`
-	${visuallyHidden};
-`;
-
-const ensureOldText = (age: string): string =>
-	age.endsWith('old') ? age : `${age} old`;
-
-const AgeWarning: React.FC<AgeWarningProps> = ({
-	age,
-	isScreenReader,
-	size = 'medium',
-}: AgeWarningProps) => {
-	const warningPrefix = 'This article is more than ';
-	const isSmall = size === 'small';
-	const ageOld = ensureOldText(age);
-
-	if (isScreenReader) {
-		return <div css={ageWarningScreenReader}>{warningPrefix + age}</div>;
-	}
-
-	return (
-		<div css={ageWarningStyles(isSmall)} aria-hidden="true">
-			<svg width="11" height="11" viewBox="0 0 11 11">
-				<path d="M5.4 0C2.4 0 0 2.4 0 5.4s2.4 5.4 5.4 5.4 5.4-2.4 5.4-5.4S8.4 0 5.4 0zm3 6.8H4.7V1.7h.7L6 5.4l2.4.6v.8z"></path>
-			</svg>
-			{warningPrefix}
-			<strong>{ageOld}</strong>
-		</div>
-	);
 };
 
 const backgroundStyles = (format: ArticleFormat): SerializedStyles => css`
